pyproject.toml
[project]
name = "snapchat-data-processing"
version = "0.1.0"
requires-python = ">=3.14"
dependencies = [
    "ipykernel>=7.1.0",
    "pillow>=12.0.0",
]
